2018.08.12 Nightly

+++ src/Board.java
@@ -1,28 +1,34 @@
 /*
  * Author: Josue Galeas
- * Last Edit: 2018.08.04
+ * Last Edit: 2018.08.12
  */
 
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.GridLayout;
+import java.util.HashMap;
 
 import javax.swing.JPanel;
 
 @SuppressWarnings("serial")
 public class Board extends JPanel
 {
-	private SideBar sideBar;
-	private ChessPanel[][] board = new ChessPanel[8][8];
+	private static SideBar sideBar;
+	private static MessageBox messageBox;
+
+	private ChessPanel[][] board;
+	private HashMap<String, String> symbolMap;
+	private String an = "EMPTY";
 	private boolean click = false;
 	private char turn = 'W';
 	private int init[] = {-1, -1};
 	private int fin[] = {-1, -1};
-	private String an = "EMPTY";
+
 	private Color borderColor = Color.DARK_GRAY;
 	private Color lightTile = Color.LIGHT_GRAY;
 	private Color darkTile = Color.GRAY;
 	private Color highlightTile = Color.GREEN;
+	private int delay = 1;
 
 	public Board()
 	{
@@ -30,6 +36,7 @@ public class Board extends JPanel
 		setPreferredSize(new Dimension(500, 500));
 		setBackground(borderColor);
 
+		board = new ChessPanel[8][8];
 		for (int x = 0; x < 8; x++)
 		{
 			for (int y = 0; y < 8; y++)
@@ -40,6 +47,20 @@ public class Board extends JPanel
 			}
 		}
 
+		symbolMap = new HashMap<String, String>();
+		symbolMap.put("WP", "♙");
+		symbolMap.put("WN", "♘");
+		symbolMap.put("WB", "♗");
+		symbolMap.put("WR", "♖");
+		symbolMap.put("WQ", "♕");
+		symbolMap.put("WK", "♔");
+		symbolMap.put("BP", "♟");
+		symbolMap.put("BN", "♞");
+		symbolMap.put("BB", "♝");
+		symbolMap.put("BR", "♜");
+		symbolMap.put("BQ", "♛");
+		symbolMap.put("BK", "♚");
+
 		createBoard();
 		updateBoard();
 		ChessPanel.setBoard(this);
@@ -49,77 +70,31 @@ public class Board extends JPanel
 		System.loadLibrary("JNI");
 	}
 
-	public native char getColor(int x, int y);
-	public native char getType(int x, int y);
 	public native void createBoard();
 	public native void deleteBoard();
+	public native char getColor(int x, int y);
+	public native char getType(int x, int y);
 	public native boolean verifyMove(int i[], int f[]);
 
-	public void setSideBar(SideBar sb) {sideBar = sb;}
+	public static void setSideBar(SideBar sb) {sideBar = sb;}
+
+	public static void setMessageBox(MessageBox mb) {messageBox = mb;}
 
 	public void setLabel(int x, int y)
 	{
-		char color = getColor(x, y);
-		char type = getType(x, y);
+		Character color = getColor(x, y);
+		Character type = getType(x, y);
+		var piece = color.toString() + type.toString();
 
-		if (color == 'E' && type == 'E')
-		{
+		if (piece.equals("EE"))
 			board[x][y].setLabel("");
-		}
-		else if (color == 'W')
-		{
-			switch (type)
-			{
-				case 'P':
-					board[x][y].setLabel("♙");
-					break;
-				case 'N':
-					board[x][y].setLabel("♘");
-					break;
-				case 'B':
-					board[x][y].setLabel("♗");
-					break;
-				case 'R':
-					board[x][y].setLabel("♖");
-					break;
-				case 'Q':
-					board[x][y].setLabel("♕");
-					break;
-				case 'K':
-					board[x][y].setLabel("♔");
-					break;
-				default:
-					board[x][y].setLabel("ERROR");
-			}
-		}
-		else if (color == 'B')
-		{
-			switch (type)
-			{
-				case 'P':
-					board[x][y].setLabel("♟");
-					break;
-				case 'N':
-					board[x][y].setLabel("♞");
-					break;
-				case 'B':
-					board[x][y].setLabel("♝");
-					break;
-				case 'R':
-					board[x][y].setLabel("♜");
-					break;
-				case 'Q':
-					board[x][y].setLabel("♛");
-					break;
-				case 'K':
-					board[x][y].setLabel("♚");
-					break;
-				default:
-					board[x][y].setLabel("ERROR");
-			}
-		}
 		else
-			board[x][y].setLabel("ERROR");
+		{
+			if (symbolMap.get(piece) != null)
+				board[x][y].setLabel(symbolMap.get(piece));
+			else
+				board[x][y].setLabel("ERROR");
+		}
 	}
 
 	public void setTile(int x, int y)
@@ -151,18 +126,30 @@ public class Board extends JPanel
 		fin[0] = fin[1] = -1;
 	}
 
-	public char flipColor(char col)
+	public char nextColor(char col)
 	{
 		if (col == 'W')
+		{
+			messageBox.setMessage("Black's Turn");
 			return 'B';
+		}
 		else if (col == 'B')
+		{
+			messageBox.setMessage("White's Turn");
 			return 'W';
+		}
 		else
+		{
+			messageBox.setMessage("ERROR");
 			return 'E';
+		}
 	}
 
 	public void newGame()
 	{
+		if (messageBox.getWaiting())
+			return;
+
 		deleteBoard();
 		createBoard();
 		updateBoard();
@@ -170,33 +157,35 @@ public class Board extends JPanel
 		if (click)
 			setTile(init[0], init[1]);
 
+		an = "EMPTY";
 		click = false;
 		turn = 'W';
 		resetInitFin();
-		an = "EMPTY";
+
+		messageBox.setMessage("White's Turn");
 		repaint();
 	}
 
 	public void processClick(int x, int y)
 	{
+		if (messageBox.getWaiting())
+			return;
+
 		if (!click)
 		{
 			char color = getColor(x, y);
 
 			if (color == 'E')
 				return;
-			else if (color != turn)
-			{
-				System.out.println("Not your piece!");
-				return;
-			}
-			else
+			else if (color == turn)
 			{
 				init[0] = x;
 				init[1] = y;
 				click = true;
 				board[x][y].setBackground(highlightTile);
 			}
+			else
+				messageBox.customMessage("Not your piece!", delay);
 		}
 		else
 		{
@@ -212,12 +201,10 @@ public class Board extends JPanel
 			{
 				updateBoard();
 				updateSideBar();
-				turn = flipColor(turn);
+				turn = nextColor(turn);
 			}
 			else
-			{
-				System.out.println("Not a valid move!");
-			}
+				messageBox.customMessage("Not a valid move!", delay);
 		}
 	}
 }
